all kinds of tiny leftovers and enable playfield reflection on default table
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2193 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ bumper.cpp
@@ -362,7 +362,7 @@ void Bumper::RenderBase(RenderDevice *pd3dDevice, const Material * const baseMat
    pd3dDevice->basicShader->SetMaterial(baseMaterial);
    pd3dDevice->basicShader->SetTexture("Texture0", &baseTexture);
    g_pplayer->m_pin3d.EnableAlphaBlend(false);
-   pd3dDevice->basicShader->SetAlphaTestValue(1.0f / 255.0f);
+   pd3dDevice->basicShader->SetAlphaTestValue((float)(1.0 / 255.0));
 
    pd3dDevice->basicShader->Begin(0);
    pd3dDevice->DrawIndexedPrimitiveVB(D3DPT_TRIANGLELIST, MY_D3DFVF_NOTEX2_VERTEX, baseVertexBuffer, 0, bumperBaseNumVertices, baseIndexBuffer, 0, bumperBaseNumFaces);
@@ -376,7 +376,7 @@ void Bumper::RenderSocket(RenderDevice *pd3dDevice, const Material * const socke
    pd3dDevice->basicShader->SetMaterial(socketMaterial);
    pd3dDevice->basicShader->SetTexture("Texture0", &socketTexture);
    g_pplayer->m_pin3d.EnableAlphaBlend(false);
-   pd3dDevice->basicShader->SetAlphaTestValue(1.0f / 255.0f);
+   pd3dDevice->basicShader->SetAlphaTestValue((float)(1.0 / 255.0));
 
    pd3dDevice->basicShader->Begin(0);
    pd3dDevice->DrawIndexedPrimitiveVB(D3DPT_TRIANGLELIST, MY_D3DFVF_NOTEX2_VERTEX, socketVertexBuffer, 0, bumperSocketNumVertices, socketIndexBuffer, 0, bumperSocketNumFaces);
@@ -390,7 +390,7 @@ void Bumper::RenderCap(RenderDevice *pd3dDevice, const Material * const capMater
    pd3dDevice->basicShader->SetMaterial(capMaterial);
    pd3dDevice->basicShader->SetTexture("Texture0", &capTexture);
    g_pplayer->m_pin3d.EnableAlphaBlend(false);
-   pd3dDevice->basicShader->SetAlphaTestValue(1.0f / 255.0f);
+   pd3dDevice->basicShader->SetAlphaTestValue((float)(1.0 / 255.0));
 
    pd3dDevice->basicShader->Begin(0);
    pd3dDevice->DrawIndexedPrimitiveVB(D3DPT_TRIANGLELIST, MY_D3DFVF_NOTEX2_VERTEX, capVertexBuffer, 0, bumperCapNumVertices, capIndexBuffer, 0, bumperCapNumFaces);
